requirements.txt
flask
flask-script
pyserial
qrcode
pillow
